+++ inject/inject-server/src/test/scala/com/twitter/inject/server/tests/EmbeddedTwitterServerIntegrationTest.scala
@@ -435,6 +435,17 @@ class EmbeddedTwitterServerIntegrationTest extends Test {
     }
   }
 
+  test("server#httpPostAdmin sends a valid POST request") {
+    val embeddedServer = new EmbeddedTwitterServer(
+      twitterServer = new TwitterServer {},
+    )
+    try {
+      embeddedServer.httpPostAdmin("/quitquitquit", andExpect = Status.Ok, withBody = "quitting\n")
+    } finally {
+      embeddedServer.close()
+    }
+  }
+
   /* utility method tests */
 
   test("method#reduceScopedFunction") {